+++ src/components/MainPage/SearchBox.js
@@ -1,27 +1,10 @@
-import React, { useState } from 'react';
-import festivalsData from '../../data/festivalsData';
-import styled from 'styled-components';
-
-const SearchForm = styled.form`
-  box-sizing: border-box;
-  width: 29rem;
-  height: 2.2rem;
-  position: relative;
-`;
-
-const SearchButt = styled.div`
-  box-sizing: border-box;
-  position: absolute;
-  size: 1rem;
-  top: 0.3rem;
-  right: 0.8rem;
-  cursor: pointer;
-`;
+import React, { useState } from "react";
+import festivalsData from "../../data/festivalsData";
 
 // ... (other imports and styled components)
 
 function SearchBox() {
-  const [searchTerm, setSearchTerm] = useState('');
+  const [searchTerm, setSearchTerm] = useState("");
   const [autocompleteResults, setAutocompleteResults] = useState([]);
   const [selectedId, setSelectedId] = useState(null);
   const [showAutocomplete, setShowAutocomplete] = useState(false);
@@ -30,7 +13,7 @@ function SearchBox() {
     const newSearchTerm = e.target.value.toLowerCase();
     setSearchTerm(newSearchTerm);
 
-    if (newSearchTerm.trim() !== '') {
+    if (newSearchTerm.trim() !== "") {
       createAutocompleteResults(newSearchTerm);
       setShowAutocomplete(true);
     } else {
@@ -70,36 +53,33 @@ function SearchBox() {
 
   return (
     <div>
-      <div style={{ position: 'relative' }}>
+      <div style={{ position: "relative" }}>
         <input
           type="text"
           value={searchTerm}
           onChange={handleInputChange}
           onFocus={() => setShowAutocomplete(true)}
           placeholder="검색어를 입력하세요..."
-          style={{ width: '300px', height: '30px' }}
+          style={{ width: "300px", height: "30px" }}
         />
         <button onClick={handleSearchClick}>검색</button>
         {showAutocomplete && autocompleteResults.length > 0 && (
           <ul
             style={{
-              background: 'red',
-              listStyleType: 'none',
+              background: "red",
+              listStyleType: "none",
               padding: 0,
-              position: 'absolute',
-              top: '40px', // Adjust the distance from the input
+              position: "absolute",
+              top: "40px", // Adjust the distance from the input
               left: 0,
-              width: '300px', // Set the width to match the input
-              maxHeight: '150px',
-              overflowY: 'auto',
+              width: "300px", // Set the width to match the input
+              maxHeight: "150px",
+              overflowY: "auto",
               zIndex: 1,
             }}
           >
             {autocompleteResults.map((result) => (
-              <li
-                key={result.id}
-                onClick={() => handleAutocompleteClick(result.id, result.name)}
-              >
+              <li key={result.id} onClick={() => handleAutocompleteClick(result.id, result.name)}>
                 {result.name}
               </li>
             ))}
@@ -109,7 +89,5 @@ function SearchBox() {
     </div>
   );
 }
-
-
 
 export default SearchBox;
